# perf: provide custom implementation for creating UUIDs.