--- a/test/common/runtime/uuid_util_test.cc
+++ b/test/common/runtime/uuid_util_test.cc
@@ -35,6 +35,14 @@ TEST(UUIDUtilsTest, mod) {
   EXPECT_EQ(7295, result);
 }
 
+TEST(UUIDUtilsTest, benchmark) {
+  Runtime::RandomGeneratorImpl random;
+
+  for (int i = 0; i < 500000; ++i) {
+    random.uuid();
+  }
+}
+
 TEST(UUIDUtilsTest, checkDistribution) {
   Runtime::RandomGeneratorImpl random;
 
@@ -42,11 +50,8 @@ TEST(UUIDUtilsTest, checkDistribution) {
   const int required_percentage = 11;
   int total_samples = 0;
   int interesting_samples = 0;
-  // For some reason, ASAN/UBSAN seems to run very slowly unless this is defined outside the loop
-  // condition.
-  const int iters = 500000;
 
-  for (int i = 0; i < iters; ++i) {
+  for (int i = 0; i < 500000; ++i) {
     std::string uuid = random.uuid();
 
     uint16_t value;