--- a/test/common/runtime/uuid_util_test.cc
+++ b/test/common/runtime/uuid_util_test.cc
@@ -47,7 +47,7 @@ TEST(UUIDUtilsTest, checkDistribution) {
     std::string uuid = random.uuid();
 
     const char c = uuid[19];
-    ASSERT_TRUE(uuid[14] == '4'); // UUID version 4 (random)
+    ASSERT_TRUE(uuid[14] == '4');                              // UUID version 4 (random)
     ASSERT_TRUE(c == '8' || c == '9' || c == 'a' || c == 'b'); // UUID variant 1 (RFC4122)
 
     uint16_t value;