This change yield ~5x performance improvement for UUID generation,
which results in ~30% performance improvement when proxying HTTP/2
traffic on localhost:

Before:
UUIDUtilsTest.benchmark:  2003 ms
h2load -c 100 -n 100000: 26467 req/s

After:
UUIDUtilsTest.benchmark:   408 ms
h2load -c 100 -n 100000: 34099 req/s

As a bonus, this implementation is fully portable and depends only
on BoringSSL's crypto.

Partially fixes #249.