@PiotrSikora 2017-08-05
Similar results on CI:

Before (current `master`):

    [ RUN      ] UUIDUtilsTest.checkDistribution
    [       OK ] UUIDUtilsTest.checkDistribution (4500 ms)

After (this PR):

    [ RUN      ] UUIDUtilsTest.benchmark
    [       OK ] UUIDUtilsTest.benchmark (795 ms)
    [ RUN      ] UUIDUtilsTest.checkDistribution
    [       OK ] UUIDUtilsTest.checkDistribution (860 ms)

@mattklein123 @htuch PTAL. Please note that ASAN test failed to build 1/146 tests because of `No space left on device`, and that's why it's marked as failed, but all tests that built passed.