@htuch 2017-08-06
(Also, if you merge with master, you should be able to pickup the `bazel.asan` workaround).