@mattklein123 2017-08-06
LGTM other than @htuch comments and my comment about disabling the benchmark test.