@PiotrSikora 2017-08-07
Bummer, the commit message was removed during merge.

Could we please not do that in the future (or more precisely, could the person merging PR make sure that the commit message is retained)? Thanks.