@mattklein123 2017-08-07
@RomanDzhabarov please see the nodes on cleaning up the commit message in https://github.com/lyft/envoy/blob/master/CONTRIBUTING.md. The goal is to clean it up but not delete it. Thanks.